Create a class Statistics that has the following functionality (the file for the class is provided in the exercise template):
a method addNumber adds a new number to the statistics

a method getCount tells the number of added numbers

the sum method tells the sum of the numbers added (the sum of an empty number statistics object is 0)

the average method tells the average of the numbers added (the average of an empty number statistics object is 0

Write a program that asks the user for numbers until the user enters -1. The program will then provide the sum of the numbers, sum of even, and odd numbers.

NOTE: Define three Statistics objects in the program