Add a new class to the project called PaymentCard create the PaymentCard object's constructor
The method eatAffordably should reduce the card's balance by € 2.60, and the method eatHeartily should reduce the card's balance by € 4.60.

Change the methods eatAffordably and eatHeartily so that they don't reduce the balance should it turn negative.

Create addMoney method further to increase the card's balance by the amount of money given as a parameter. However, the card's balance may not exceed 150 euros. If there is an attempt to top it up with a negative amount, the value on the card will not change.

Write code in the main method of the MainProgram class that contains the following sequence of events:

Create Paul's card. The opening balance of the card is 20 euros.
Create Matt's card. The opening balance of the card is 30 euros.
Paul eats heartily.
Matt eats affordably.
The cards' values are printed .
Paul tops up 20 euros.
Matt eats heartily.
The cards' value are printed .
Paul eats affordably.
Paul eats affordably.
Matt tops up 50 euros.
The cards' value are printed.