Create a class named HealthStation and  a class named Person. Create a method 'weigh' for the HealthStation that receives a person as a parameter, and it is meant to return to its caller the weight of that person. The weight information can be found by calling a suitable method of the person person.
Modify the state of the object that is received as a parameter. Write a method called public void feed(Person person) for the health station. It should increase the weight of the parameter person by one.

Create a new method called public int weighings() for the health station. It should tell how many weighings the health station has performed.