An improved solution is to make the cards dumb; unaware of the prices and products that are sold, and only keeping track of their balance. All the intelligence is better placed in separate objects, payment terminals.
Let's first implement the dumb version of the PaymentCard. The card only has methods for asking for the balance, adding money, and taking money.

First, let's create a terminal that's suitable for cash payments. The terminal starts with 1000 euros in it.

Let's extend our payment terminal to also support card payments. We are going to create new methods for the terminal. It receives a payment card as a parameter, and decreases its balance by the price of the meal that was purchased. Card payments don't increase the amount of cash in the register.

Let's create a method for the terminal that can be used to add money to a payment card. Recall that the payment that is received when adding money to the card is stored in the register.