Your task is to create a few methods that can be used to compare apartments that are being sold.
Comparing sizes

Create a method public boolean largerThan(Apartment compared) that returns true if the apartment object whose method is called has a larger total area than the apartment object that is being compared.

Price difference

Create a method public int priceDifference(Apartment compared) that returns the price difference of the apartment object whose method was called and the apartment object received as the parameter. The price difference is the absolute value of the difference of the prices

More expensive?

Write a method public boolean moreExpensiveThan(Apartment compared) that returns true if the apartment object whose method is called is more expensive than the apartment object being compared.